requirements.txt
python-dotenv>=0.20.0
google-generativeai>=0.5.0
Pillow>=9.0.0